requirement.txt
absl-py==0.9.0
altair==4.1.0
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.13
botocore==1.17.13
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
future==0.18.2
gast==0.2.2
gluoncv==0.7.0
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.30.0
h5py==2.10.0
idna==2.6
imageio==2.8.0
importlib-metadata==1.7.0
ipykernel==5.3.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.5
jupyter-core==4.6.3
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mxnet==1.6.0
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
notebook==6.0.3
numpy==1.16.6
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
portalocker==1.7.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.4.0b2
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.18.4
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
streamlit==0.58.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
tqdm==4.47.0
traitlets==4.3.3
tzlocal==2.1
urllib3==1.22
validators==0.15.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
